backend/requirements.txt
# Web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Geo
geoalchemy2==0.14.3
h3==3.7.7
shapely==2.0.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Cache / Queue
redis==5.0.1
celery[redis]==5.3.6
flower==2.0.1

# ML
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4

# Utils
httpx==0.27.0
python-dateutil==2.8.2
orjson==3.9.13

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0

# Rate limiting
slowapi==0.1.9
